+++ demisto_sdk/commands/content_graph/objects/indicator_field.py
@@ -1,3 +1,4 @@
+from pathlib import Path
 from typing import Optional
 
 from pydantic import Field
@@ -11,6 +12,7 @@ from demisto_sdk.commands.content_graph.objects.indicator_incident_field import 
 
 class IndicatorField(IndicatorIncidentField, content_type=ContentType.INDICATOR_FIELD):  # type: ignore[call-arg]
     associated_to_all: bool = Field(alias="associatedToAll")
+    type: str
 
     def summary(
         self,
@@ -20,3 +22,11 @@ class IndicatorField(IndicatorIncidentField, content_type=ContentType.INDICATOR_
         summary = super().summary(marketplace, incident_to_alert)
         summary["id"] = f"indicator_{self.object_id}"
         return summary
+
+    @staticmethod
+    def match(_dict: dict, path: Path) -> bool:
+        if "id" in _dict:
+            _id = _dict["id"].lower()
+            if _id.startswith("indicator"):
+                return True
+        return False