+++ demisto_sdk/commands/content_graph/objects/list.py
@@ -8,6 +8,7 @@ from demisto_sdk.commands.common.constants import MarketplaceVersions
 from demisto_sdk.commands.common.handlers import JSON_Handler
 from demisto_sdk.commands.content_graph.common import ContentType
 from demisto_sdk.commands.content_graph.objects.content_item import ContentItem
+from demisto_sdk.commands.prepare_content.list_unifier import ListUnifier
 
 json = JSON_Handler()
 logger = logging.getLogger("demisto-sdk")
@@ -15,6 +16,7 @@ logger = logging.getLogger("demisto-sdk")
 
 class List(ContentItem, content_type=ContentType.LIST):  # type: ignore[call-arg]
     type: str
+    is_unified: bool
 
     def _upload(
         self,
@@ -31,3 +33,27 @@ class List(ContentItem, content_type=ContentType.LIST):  # type: ignore[call-arg
                 body=json.loads((dir_path / self.normalize_name).read_text()),
                 response_type="object",
             )
+
+    def prepare_for_upload(
+        self,
+        current_marketplace: MarketplaceVersions = MarketplaceVersions.XSOAR,
+        **kwargs
+    ) -> dict:
+        data = (
+            self.data
+            if kwargs.get("unify_only")
+            else super().prepare_for_upload(current_marketplace)
+        )
+        if self.is_unified:
+            return data
+        return ListUnifier.unify(self.path, data, marketplace=current_marketplace)
+
+    @staticmethod
+    def match(_dict: dict, path: Path) -> bool:
+        if (
+            isinstance(_dict, dict)
+            and {"data", "allRead", "truncated"}.intersection(_dict.keys())
+            and path.suffix == ".json"
+        ):
+            return True
+        return False