+++ demisto_sdk/commands/pre_commit/hooks/ruff.py
@@ -1,6 +1,6 @@
+import os
 from copy import deepcopy
-from pathlib import Path
-from typing import Any, Dict, Set
+from typing import Any, Dict
 
 from demisto_sdk.commands.pre_commit.hooks.hook import (
     Hook,
@@ -23,18 +23,13 @@ class RuffHook(Hook):
 
     def prepare_hook(
         self,
-        python_version_to_files: Dict[str, Set[Path]],
-        github_actions: bool = False,
-        **kwargs,
     ) -> None:
         """
         Prepares the Ruff hook for each Python version.
         Changes the hook's name, files and the "--target-version" argument according to the Python version.
         Args:
-            python_version_to_files (Dict[str, Set[Path]]): dictionary mapping python version to files
-            github_actions (bool, optional): Whether to use github actions format. Defaults to False.
         """
-        for python_version in python_version_to_files:
+        for python_version in self.context.python_version_to_files:
             hook: Dict[str, Any] = {
                 "name": f"ruff-py{python_version}",
             }
@@ -43,8 +38,14 @@ class RuffHook(Hook):
                 f"--target-version={self._python_version_to_ruff(python_version)}"
             )
             safe_update_hook_args(hook, target_version)
-            if github_actions:
+            if os.getenv("GITHUB_ACTIONS", False):
                 hook["args"].append("--format=github")
-            hook["files"] = join_files(python_version_to_files[python_version])
+            hook["files"] = join_files(
+                {
+                    file
+                    for file in self.context.python_version_to_files[python_version]
+                    if file.suffix == ".py"
+                }
+            )
 
             self.hooks.append(hook)