+++ demisto_sdk/scripts/merge_pytest_reports.py
@@ -8,7 +8,6 @@ from junitparser import JUnitXml
 
 from demisto_sdk.commands.common.content_constant_paths import CONTENT_PATH
 from demisto_sdk.commands.common.logger import logger
-from demisto_sdk.commands.coverage_analyze.helpers import coverage_files
 
 
 def fix_coverage_report_path(coverage_file: Path) -> bool:
@@ -35,10 +34,10 @@ def fix_coverage_report_path(coverage_file: Path) -> bool:
                         cursor.execute(
                             "DELETE FROM file WHERE id = ?", (id_,)
                         )  # delete the file from the coverage report, as it is not relevant.
-                    if not file.startswith("/content"):
+                    if not file.startswith("/src"):
                         # means that the .coverage file is already fixed
                         continue
-                    file = Path(file).relative_to("/content")
+                    file = Path(file).relative_to("/src")
                     if (
                         not (CONTENT_PATH / file).exists()
                         or file.parent.name
@@ -59,8 +58,8 @@ def fix_coverage_report_path(coverage_file: Path) -> bool:
             shutil.copy(temp_file.name, coverage_file)
             return True
     except Exception:
-        logger.warning(f"Broken .coverage file found: {file}, deleting it")
-        file.unlink(missing_ok=True)
+        logger.warning(f"Broken .coverage file found: {coverage_file}, deleting it")
+        coverage_file.unlink(missing_ok=True)
         return False
 
 
@@ -68,10 +67,11 @@ def merge_coverage_report():
     coverage_path = CONTENT_PATH / ".coverage"
     coverage_path.unlink(missing_ok=True)
     cov = coverage.Coverage(data_file=coverage_path)
-    if not (files := coverage_files()):
+    coverage_paths = CONTENT_PATH / ".pre-commit" / "coverage"
+    if not (files := list(coverage_paths.iterdir())):
         logger.warning("No coverage files found, skipping coverage report.")
         return
-    fixed_files = [file for file in files if fix_coverage_report_path(Path(file))]
+    fixed_files = [str(file) for file in files if fix_coverage_report_path(Path(file))]
     cov.combine(fixed_files)
     cov.xml_report(outfile=str(CONTENT_PATH / "coverage.xml"))
     for file in files:
@@ -80,7 +80,9 @@ def merge_coverage_report():
 
 
 def merge_junit_reports():
-    report_files = CONTENT_PATH.rglob(".report_pytest.xml")
+    junit_reports_path = CONTENT_PATH / ".pre-commit" / "pytest-junit"
+
+    report_files = junit_reports_path.iterdir()
     if reports := [JUnitXml.fromfile(str(file)) for file in report_files]:
         report = reports[0]
         for rep in reports[1:]: